Migrate code to PlayerGroundMotor
Migrates code responsible for enabling input to move the player along the ground to PlayerGroundMotor.  Also migrates the code enabling moving platforms to move the player.
