+++ Assets/Scripts/Game/Player/PlayerGroundMotor.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 73032437e0d00e246b52351a1e1f70c6
+timeCreated: 1526254620
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
